requirements_full.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
attrs==25.4.0
-e git+https://github.com/anju2246/AutoEditingSFC.git@7a82140bd337dc35d606f8935f96a5817bcf64c4#egg=autoeditingsfc
autoflake==2.3.1
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deep-sort-realtime==1.3.2
demucs==4.0.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
dora_search==0.1.12
einops==0.8.1
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-genai==1.59.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.2
hyperframe==6.1.0
HyperPyYAML==1.2.3
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.31
mlx==0.30.3
mlx-metal==0.30.3
mlx-whisper==0.4.3
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.15.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
openunmix==1.3.0
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
postgrest==2.27.2
primePy==1.3
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannote.audio==3.1.1
pyannoteai-sdk==0.4.0
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.45
starlette==0.50.0
storage3==2.27.2
streamlit==1.53.0
StrEnum==0.4.15
strictyaml==1.7.3
submitit==1.5.4
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.9.1
torchcodec==0.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
treetable==0.2.6
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
